+++ one_compliance/one_compliance/doctype/task_bulk_assignment/task_bulk_assignment.py
@@ -7,7 +7,7 @@ import json
 from frappe import _
 from frappe.desk.form.assign_to import add as add_assignment
 
-class TaskReassignment(Document):
+class TaskBulkAssignment(Document):
 	@frappe.whitelist()
 	def get_allocation_entries(self):
 		self.prepare_task_reassign()
@@ -254,146 +254,3 @@ def allocate_tasks_to_employee(selected_task_ids, selected_employee_ids):
 
 	# Return a success message
 	return "Tasks allocated successfully"
-
-@frappe.whitelist()
-def get_compliance_categories_for_user(doctype, txt, searchfield, start, page_len, filters):
-    user_id = filters.get('user_id')
-    print(user_id)
-
-    # Find the Employee based on the user_id
-    employee = frappe.get_doc("Employee", {"user_id": user_id})
-
-    if employee:
-        # Get the department of the employee
-        department = employee.department
-
-        # Query the Compliance Category DocType
-        compliance_categories = frappe.get_all("Compliance Category", filters={"department": department}, fields=["name", "department"])
-
-        category_names = []  # Initialize an empty list to store user information
-
-        for category in compliance_categories:
-            name = category['name']
-            department = category['department']
-
-            category_names.append((name, department))  # Add email and full name as a tuple to the list
-
-        return category_names  # Return the list of email IDs and full names as tuples
-    return []
-
-import frappe
-
-@frappe.whitelist()
-def get_compliance_executives(compliance_category):
-    child_table_data = frappe.get_all(
-        'Compliance Executive',
-        filters={'parent': compliance_category},
-        fields=['employee', 'designation', 'employee_name']
-    )
-    return child_table_data
-
-@frappe.whitelist()
-def add_employee_to_compliance_executive(employee, compliance_category):
-    try:
-        # Check if the employee and compliance category exist
-        employee_doc = frappe.get_doc("Employee", {"user_id": employee})
-        compliance_category_doc = frappe.get_doc("Compliance Category", compliance_category)
-
-        if employee_doc and compliance_category_doc:
-            # Check if the employee is already in the Compliance Executive table for the category
-            existing_executive = None
-            for executive in compliance_category_doc.compliance_executive:
-                if executive.employee == employee_doc.name:
-                    existing_executive = executive
-                    break
-
-            if not existing_executive:
-                # Create a new "Compliance Executive" document and set its fields
-                executive = frappe.new_doc("Compliance Executive")
-                executive.employee = employee_doc.name
-                executive.designation = employee_doc.designation
-                executive.employee_name = employee_doc.employee_name
-
-                # Append the new executive to the Compliance Executive table
-                compliance_category_doc.append("compliance_executive", executive)
-
-                # Save the Compliance Category document with the updated table
-                compliance_category_doc.save()
-
-                frappe.db.commit()
-                return True
-            else:
-                frappe.msgprint("Employee is already existing")
-                return False  # Employee is already in the table
-        else:
-            return False  # Employee or Compliance Category does not exist
-    except Exception as e:
-        frappe.log_error(frappe.get_traceback(), _("Error in adding employee to Compliance Executive"))
-        return False
-
-@frappe.whitelist()
-def get_available_subcategories(compliance_category, employee):
-    # Query the database to retrieve subcategories related to the selected category
-    subcategories = frappe.get_all(
-        "Compliance Sub Category",
-        filters={"compliance_category": compliance_category},
-        fields=["name"]
-    )
-
-    # Check if the employee exists in the compliance executive of each subcategory
-    employee_doc = frappe.get_doc("Employee", {"user_id": employee})
-    for subcategory in subcategories:
-        compliance_subcategory_doc = frappe.get_doc("Compliance Sub Category", subcategory["name"])
-
-        is_added = False
-        for executive in compliance_subcategory_doc.compliance_executive:
-            if executive.employee == employee_doc.name:
-                is_added = True
-                break
-
-        subcategory["status"] = "added" if is_added else "not added"
-
-    return subcategories
-
-@frappe.whitelist()
-def add_to_subcategories(employee, compliance_category, selected_subcategories):
-
-    try:
-        # Check if the employee and compliance category exist
-        employee_doc = frappe.get_doc("Employee", {"user_id": employee})
-        compliance_category_doc = frappe.get_doc("Compliance Category", compliance_category)
-        compliance_subcategories = json.loads(selected_subcategories)
-
-        if employee_doc and compliance_category_doc:
-            for subcategory in compliance_subcategories:
-
-                compliance_subcategory_doc = frappe.get_doc("Compliance Sub Category", {"name": subcategory})
-
-                if compliance_subcategory_doc:
-
-                    existing_executive = None
-                    for executive in compliance_subcategory_doc.compliance_executive:
-                        if executive.employee == employee_doc.name:
-                            existing_executive = executive
-                            break
-
-                    if not existing_executive:
-                        # Create a new "Compliance Executive" document and set its fields
-                        executive = frappe.new_doc("Compliance Executive")
-                        executive.employee = employee_doc.name
-                        executive.designation = employee_doc.designation
-                        executive.employee_name = employee_doc.employee_name
-
-                        # Append the new executive to the Compliance Executive table
-                        compliance_subcategory_doc.append("compliance_executive", executive)
-
-                        # Save the Compliance Sub Category document with the updated table
-                        compliance_subcategory_doc.save()
-
-            frappe.db.commit()
-            return True
-        else:
-            return False
-    except Exception as e:
-        frappe.log_error(frappe.get_traceback(), _("Error in adding employee to Compliance Sub Categories"))
-        return False
