+++ one_compliance/one_compliance/doctype/task_bulk_assignment/test_task_bulk_assignment.py
@@ -5,5 +5,5 @@
 from frappe.tests.utils import FrappeTestCase
 
 
-class TestTaskReassignment(FrappeTestCase):
+class TestTaskBulkAssignment(FrappeTestCase):
 	pass
